% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{canterbury}
\alias{canterbury}
\title{Polygon for the Canterbury region of New Zealand}
\value{
A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(canterbury)
}
\keyword{datasets}
