% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uk_shape.R
\name{get_uk_shape}
\alias{get_uk_shape}
\title{Create a local copy of uk shape file}
\usage{
get_uk_shape()
}
\value{
logical TRUE if the operation was successful
}
\description{
This function will attempt to create a local copy of the uk shape file
such that you can make sure that you are happy with reading in data from shape files
}
