% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{meth}
\alias{meth}
\title{Locations of meth lab busts in the US}
\value{
A tibble or data frame.
}
\description{
Locations of meth lab busts in the US
}
\examples{
data(meth)
}
\keyword{datasets}
