% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{nz}
\alias{nz}
\title{Polygons for the regions of New Zealand}
\value{
A tibble or data frame.

A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source

See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(nz)
data(nz)
}
\keyword{datasets}
