% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{nz_elev}
\alias{nz_elev}
\title{Elevation data for New Zealand stored as raster}
\value{
A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(nz_elev)
}
\keyword{datasets}
