% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{nz_missing}
\alias{nz_missing}
\title{Polygons for the regions of New Zealand, excluding Canterbury}
\value{
A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(nz_missing)
}
\keyword{datasets}
