% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{ukdata}
\alias{ukdata}
\title{Data for the regions of the UK}
\value{
A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(ukdata)
}
\keyword{datasets}
