% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{ukhills}
\alias{ukhills}
\title{Locations of hills in the UK}
\value{
A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(ukhills)
}
\keyword{datasets}
