% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{ukhills_big}
\alias{ukhills_big}
\title{Location of hills in the UK}
\value{
A tibble or data frame.
}
\description{
Locations of the largest hill in each region of the UK.
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(ukhills_big)
}
\keyword{datasets}
