% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{worldstats}
\alias{worldstats}
\title{Statistics for each country in the world}
\value{
A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(worldstats)
}
\keyword{datasets}
