% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_files.R
\docType{data}
\name{wonders}
\alias{wonders}
\title{Latitude and Longitude for the seven wonders of the modern world}
\value{
A tibble or data frame.
}
\description{
See vignette("reading", package = "jrSpatial") for
source
}
\examples{
data(nz_missing)
}
\keyword{datasets}
